+++ database/migrations/2024_12_30_164923_create_property_amenities_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('property_amenities')) {
+            return;
+        }
+
         Schema::create('property_amenities', function (Blueprint $table) {
             $table->id();
             $table->foreignId('property_id')->constrained('properties')->onDelete('cascade');
